# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MEI Fácil API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /taxes:
    get:
      operationId: listTaxes
      tags: [taxes]
      summary: List DAS tax payments
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxPayment"
    post:
      operationId: createTax
      tags: [taxes]
      summary: Create a DAS tax record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxPayment"
  /taxes/{id}/pay:
    patch:
      operationId: markTaxPaid
      tags: [taxes]
      summary: Mark a DAS tax as paid
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxPaymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxPayment"
  /taxes/simulate-penalty:
    post:
      operationId: simulatePenalty
      tags: [taxes]
      summary: Simulate DAS late payment penalty
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PenaltySimulationInput"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PenaltySimulation"
  /taxes/next-due:
    get:
      operationId: getNextDueTax
      tags: [taxes]
      summary: Get the next due DAS payment
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxPayment"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TaxPayment:
      type: object
      required: [id, userId, competence, dueDate, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        competence:
          type: string
        dueDate:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, overdue]
        paidAt:
          type: ["string", "null"]
        daysUntilDue:
          type: ["integer", "null"]
        createdAt:
          type: string
    TaxInput:
      type: object
      required: [competence, dueDate, amount]
      properties:
        competence:
          type: string
        dueDate:
          type: string
        amount:
          type: number
    TaxPaymentUpdate:
      type: object
      properties:
        paidAt:
          type: ["string", "null"]
    PenaltySimulationInput:
      type: object
      required: [amount, daysLate]
      properties:
        amount:
          type: number
        daysLate:
          type: integer
    PenaltySimulation:
      type: object
      required: [originalAmount, fine, interest, total, daysLate]
      properties:
        originalAmount:
          type: number
        fine:
          type: number
        interest:
          type: number
        total:
          type: number
        daysLate:
          type: integer
...