# lib/api-zod/src/generated/types/penaltySimulation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MEI Fácil API
 * OpenAPI spec version: 0.1.0
 */

export interface PenaltySimulation {
  originalAmount: number;
  fine: number;
  interest: number;
  total: number;
  daysLate: number;
}